refactor: move mesh screenshot helper
