+++ glacium/post/analysis/__init__.py
@@ -1,5 +1,6 @@
 from .ice_thickness import read_wall_zone, process_wall_zone, plot_ice_thickness
 from .ice_contours import load_contours, plot_overlay, animate_growth
+from .mesh_screenshots import generate_wireframes
 
 __all__ = [
     "read_wall_zone",
@@ -8,4 +9,5 @@ __all__ = [
     "load_contours",
     "plot_overlay",
     "animate_growth",
+    "generate_wireframes",
 ]
